criação lógica deleção de paciente

+++ LabMedicineAPI/Service/Paciente/PacienteServices.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using LabMedicineAPI.DTOs.Endereco;
 using LabMedicineAPI.DTOs.Paciente;
 using LabMedicineAPI.Model;
 using LabMedicineAPI.Repositories;
@@ -22,7 +23,7 @@ namespace LabMedicineAPI.Service.Paciente
         public IEnumerable<PacienteGetDTO> Get()
         {
             IEnumerable<PacienteModel> pacientes = _repository.GetAll();
-            IEnumerable<PacienteGetDTO>pacientesDTO = _mapper.Map<IEnumerable<PacienteGetDTO>>(pacientes);
+            IEnumerable<PacienteGetDTO> pacientesDTO = _mapper.Map<IEnumerable<PacienteGetDTO>>(pacientes);
             return pacientesDTO;
         }
 
@@ -38,13 +39,14 @@ namespace LabMedicineAPI.Service.Paciente
             var paciente = _mapper.Map<PacienteModel>(pacienteEnderecoCreateDTO.Paciente);
             var enderecoModel = _mapper.Map<EnderecoModel>(pacienteEnderecoCreateDTO.Endereco);
 
-            paciente.Endereco = enderecoModel;
-
+            paciente = _repository.GetAll().Where(x => x.CPF == pacienteEnderecoCreateDTO.Paciente.CPF).FirstOrDefault();
             _repository.Create(paciente);
+
+            paciente.Endereco = enderecoModel;
             _enderecoRepository.Create(enderecoModel);
 
             return pacienteEnderecoCreateDTO;
-            
+
 
             //PacienteModel paciente = _mapper.Map<PacienteEnderecoCreateModel>(pacienteCreateDTO);
             //_repository.Create(paciente);
@@ -55,21 +57,76 @@ namespace LabMedicineAPI.Service.Paciente
             //return pacienteGet;
 
         }
-        public PacienteGetDTO PacienteUpdateDTO(int id, PacienteUpdateDTO updatePacienteDTO)
+        public PacienteEnderecoUpdateDTO PacienteEnderecoUpdate(int id, PacienteEnderecoUpdateDTO pacienteEnderecoUpdateDTO)
         {
             PacienteModel paciente = _repository.GetById(id);
-            paciente = _mapper.Map(updatePacienteDTO, paciente);
+
+            if (paciente != null)
+            {
+                _mapper.Map(pacienteEnderecoUpdateDTO.Paciente, paciente); // Mapeia os dados do paciente do DTO para o modelo do paciente.
+            }
+
+            EnderecoModel endereco = _mapper.Map<EnderecoModel>(pacienteEnderecoUpdateDTO.Endereco);
+            paciente.Endereco = endereco;
 
             _repository.Update(paciente);
-            PacienteModel pacienteUpdated = _repository.GetById(id);
-            PacienteGetDTO pacienteGet = _mapper.Map<PacienteGetDTO>(pacienteUpdated);
-            return pacienteGet;
+
+            PacienteEnderecoUpdateDTO pacienteAtualizadoDTO = new PacienteEnderecoUpdateDTO
+            {
+                Paciente = _mapper.Map<PacienteUpdateDTO>(paciente),
+                Endereco = _mapper.Map<EnderecoUpdateDTO>(endereco)
+            };
+
+            return pacienteAtualizadoDTO;
         }
 
         public bool DeletePaciente(int id)
         {
-            return _repository.Delete(id);
+            var paciente = _repository.GetById(id);
+
+            if (paciente != null)
+            {
+                var impedimento = VerificarImpedimentosDelecao(paciente);
+                if (paciente.Endereco != null)
+                {
+                    _enderecoRepository.Delete(paciente.Endereco.Id);
+                }
+
+                return _repository.Delete(id);
+            }
+
+            return false;
+        }
+        private string VerificarImpedimentosDelecao(PacienteModel paciente)
+        {
+            if (paciente.Consultas != null && paciente.Consultas.Count > 0)
+            {
+                return "Não é possível excluir o paciente, pois ele tem consultas vinculadas.";
+            }
+
+            if (paciente.Dietas != null && paciente.Dietas.Count > 0)
+            {
+                return "Não é possível excluir o paciente, pois ele tem dietas vinculadas.";
+            }
+
+            if (paciente.Exames != null && paciente.Exames.Count > 0)
+            {
+                return "Não é possível excluir o paciente, pois ele tem exames vinculados.";
+            }
+
+            if (paciente.Exercicios != null && paciente.Exercicios.Count > 0)
+            {
+                return "Não é possível excluir o paciente, pois ele tem exercícios vinculados.";
+            }
+
+            if (paciente.Medicamentos != null && paciente.Medicamentos.Count > 0)
+            {
+                return "Não é possível excluir o paciente, pois ele tem medicamentos vinculados.";
+            }
+
+            return null; 
         }
 
     }
+
 }
